requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
aiortc==1.9.0
aiohttp
python-socketio[client]==5.11.2
av==11.0.0